refactor : 총 가격 계산 PriceSection 컴포넌트에서 수행하도록 변경
